@@ -10,12 +10,12 @@
   },
   "dependencies": {
     "core-js": "^3.15.1",
-    "element-ui": "^2.15.2",
-    "vue": "^2.6.14",
-    "vue-class-component": "^7.2.6",
+    "element-plus": "^1.0.2-beta.53",
+    "vue": "^3.0.5",
+    "vue-class-component": "^8.0.0-beta.1",
     "vue-clipboard2": "^0.3.1",
     "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.5.2"
+    "vue-router": "^4.0.0-alpha.6"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^4.28.0",
@@ -25,39 +25,17 @@
     "@vue/cli-plugin-router": "~4.5.13",
     "@vue/cli-plugin-typescript": "~4.5.13",
     "@vue/cli-service": "~4.5.13",
+    "@vue/compiler-sfc": "^3.1.2",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^7.0.0",
     "eslint": "^7.29.0",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-standard": "^4.1.0",
     "eslint-plugin-vue": "^7.12.1",
     "prettier": "^2.3.2",
-    "typescript": "~4.3.4",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/recommended",
-      "eslint:recommended",
-      "@vue/typescript",
-      "@vue/prettier",
-      "@vue/prettier/@typescript-eslint"
-    ],
-    "parserOptions": {
-      "ecmaVersion": 2020
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "typescript": "~4.3.4"
+  }
 }
