@@ -1,8 +1,7 @@
 <template>
   <div id="app">
     <div id="nav" class="navBar">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link> |
+      <router-link to="/">About</router-link> |
       <router-link to="/todo-list">Todo List</router-link> |
       <router-link to="/counter">Counter</router-link> |
       <router-link to="/save-links">Save Links</router-link>
@@ -10,29 +9,6 @@
     <router-view />
   </div>
 </template>
-
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-}
-
-#nav {
-  padding: 30px;
-}
-
-#nav a {
-  font-weight: bold;
-  color: #2c3e50;
-}
-
-#nav a.router-link-exact-active {
-  color: #42b983;
-}
-</style>
 
 <style scoped>
 .navBar {
