@@ -1,17 +1,8 @@
-import Vue from 'vue'
-import VueRouter, { RouteConfig } from 'vue-router'
-import Home from '../views/Home.vue'
+import { createRouter, createWebHashHistory } from 'vue-router'
 
-Vue.use(VueRouter)
-
-const routes: Array<RouteConfig> = [
+const routes: Array<any> = [
   {
     path: '/',
-    name: 'Home',
-    component: Home,
-  },
-  {
-    path: '/about',
     name: 'About',
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
@@ -26,7 +17,7 @@ const routes: Array<RouteConfig> = [
   },
   {
     path: '/counter',
-    name: 'Counetr',
+    name: 'Counter',
     component: () => import('@/views/Counter.vue'),
   },
   {
@@ -36,7 +27,8 @@ const routes: Array<RouteConfig> = [
   },
 ]
 
-const router = new VueRouter({
+const router = createRouter({
+  history: createWebHashHistory(),
   routes,
 })
 
